[misc] size-change signal signature uses gints
During a short periode it used gfloats for the texture size, but got
reverted to gints shortly after. I forgot to update the signal
signature.

+++ examples/video-player.c
@@ -224,8 +224,8 @@ input_cb (ClutterStage *stage,
 
 static void
 size_change (ClutterTexture *texture, 
-             gfloat          width,
-             gfloat          height,
+             gint            width,
+             gint            height,
              VideoApp       *app)
 {
   ClutterActor *stage = app->stage;
