+++ examples/video-sink.c
@@ -28,8 +28,8 @@
 
 void
 size_change (ClutterTexture *texture,
-             gfloat          width,
-             gfloat          height,
+             gint            width,
+             gint            height,
              gpointer        user_data)
 {
   ClutterActor *stage;
