+++ tests/test-yuv-upload.c
@@ -62,8 +62,8 @@ parse_fourcc (const gchar *fourcc)
 
 void
 size_change (ClutterTexture *texture,
-             gfloat          width,
-             gfloat          height,
+             gint            width,
+             gint            height,
              gpointer        user_data)
 {
   ClutterActor *stage;
